backend/requirements.txt
flask==2.2.5
flask_cors==4.0.0
flask-sqlalchemy==3.0.5
Werkzeug==2.2.3
robotframework==6.1.1